requirements.txt
PyYAML==3.11
docopt==0.6.2
nose==1.3.7
pep8==1.7.0
psutil==3.4.2
redis==2.10.5
requests==2.9.1
